--- data/appianFileUploadNoCORS
@@ -1,60 +1,0 @@
-async function uploadFileNoCORS(file) {
-  function createFormData(file) {
-    const formData = new FormData();
-    formData.append("file", file); // Assuming your API expects the file under the key 'file'
-    return formData;
-  }
-  try {
-    const myHeaders = new Headers();
-    myHeaders.append("Content-Type", "multipart/form-data");
-    myHeaders.append("Appian-Document-Name", "name");
-    myHeaders.append(
-      "Appian-API-Key",
-      "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI3Y2UyYzZjOS0xYjk2LTRmMDMtYWE5OC02YzlmMjI0Mjc1NmUifQ.1RgP9FVjKnIxkIH3WMa9WEkgNx32oUnNSVnv4az0geo"
-    );
-    // myHeaders.append("Access-Control-Allow-Origin", "*");
-    console.log(myHeaders);
-    console.log("file" + file);
-    // console.log("event body" + event.body);
-    // console.log("event body 2" + event['body']);
-    // console.log("event stringy" + JSON.stringify(event));
-
-    const uploadRequestOptions = {
-      method: "POST",
-      headers: myHeaders,
-      body: createFormData(file),
-      redirect: "follow",
-      mode: "no-cors",
-    };
-    
-    console.log(uploadRequestOptions);
-
-    const uploadResponse = await fetch(
-      "https://vassuk.appiancloud.com/suite/webapi/upload-document",
-      uploadRequestOptions
-    );
-    const appianResponseText = await uploadResponse.text();
-    console.log(uploadResponse);
-    console.log("Appian Response Status:", uploadResponse.status);
-    console.log("Appian Response Text:", appianResponseText);
-
-    return {
-      statusCode: 200, // Or adjust based on Appian Response
-      body: appianResponseText,
-    };
-  } catch (error) {
-    console.error(error);
-    return {
-      statusCode: 500,
-      headers: {
-        "Access-Control-Allow-Origin": "*",
-        "Access-Control-Allow-Methods": "OPTIONS,POST",
-        "Access-Control-Allow-Headers":
-          "Content-Type, Authorization, appian-document-name",
-      },
-      body: "Error uploading to Appian",
-    };
-  }
-}
-
-export default uploadFileNoCORS;
