--- data/callCamundaWebhook.js
@@ -1,27 +1,0 @@
-//------------------------------------------------------------------------------------------------------------------------------//
-//-------------------------------------------------------EDIT BELOW - 1---------------------------------------------------------//
-//------------------------------------------------------------------------------------------------------------------------------//
-//Replace null return with real fetch request-----------------------------------------------------------------------------------//
-//Use your created API URL------------------------------------------------------------------------------------------------------//
-//------------------------------------------------------------------------------------------------------------------------------//
-export async function callCamundaWebhook(data) {
-  return null;
-  /**
-  try {
-    const response = await fetch(
-      <your fetch setup>
-    ).catch(error => {
-      throw new Error(error);
-    });
-    const result = await response.text();
-    console.log(result);
-    return result; // Return the result for further use
-  } catch (error) {
-    console.error(error);
-    throw error; // Rethrow the error for potential handling elsewhere
-  }
-  **/
-}
-//------------------------------------------------------------------------------------------------------------------------------//
-//-------------------------------------------------------EDIT ABOVE - 1---------------------------------------------------------//
-//------------------------------------------------------------------------------------------------------------------------------//
