+++ src/functions/apis.js
@@ -98,3 +98,38 @@ function getHardCodedLookupValues(tableName) {
       return []
   }
 }
+
+
+//------------------------------------------------------------------------------------------------------------------------------//
+//-------------------------------------------------------EDIT BELOW - 5---------------------------------------------------------//
+//------------------------------------------------------------------------------------------------------------------------------//
+//Replace null return with real fetch request-----------------------------------------------------------------------------------//
+//Use your created API URL------------------------------------------------------------------------------------------------------//
+//------------------------------------------------------------------------------------------------------------------------------//
+export async function callCamundaWebhook(data) {
+  // return null;
+  
+  try {
+    const response = await fetch(
+      `https://f9r92naghh.execute-api.eu-west-2.amazonaws.com/v1/arashLoanProcessing`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json"
+        },
+        body: JSON.stringify(data)
+      }
+    ).catch(error => {
+      throw new Error(error);
+    });
+    const result = await response.text();
+    console.log(result);
+    return result; // Return the result for further use
+  } catch (error) {
+    console.error(error);
+    throw error; // Rethrow the error for potential handling elsewhere
+  }
+  
+}
+//------------------------------------------------------------------------------------------------------------------------------//
+//-------------------------------------------------------EDIT ABOVE - 5---------------------------------------------------------//
+//------------------------------------------------------------------------------------------------------------------------------//
